feat: move flashcard viewer into separate component
